Raster: split canvas.rs into blit/gradient/skia modules behind a DrawTarget
Extract the mask sampling types, tiny-skia fast paths, free blitters, and
gradient tile fast paths out of the 2193-line canvas.rs into canvas/{mask,
skia,blit,gradient}.rs. Introduce a DrawTarget borrow-view that owns the
pixmap, combined mask, and buffer pool so mask materialization derives its
size from the target pixmap instead of a threaded content size.
